Refactor the code structure
